initial cygwin support
Add cygwin as a supported platform, for FSAL_POSIX only.  The
change also depends on tirpc.  The support is fairly complete.
Getmntent is serialized.  There is no quota support.  The largest
visible gap is in netdb functions.  A likely approach exists to
resolve this, with some external library improvement.

Two changes required for Cygwin are general.  First, one unused
pthread mutexattr and the nop mutexattr and condattr values used
in rw_lock_init are elided.  There is a bug in Cygwin here, but
the intent is to initialize with default parameters, so the code
has no real effect.  Next, there is a narrowing typecast in the
insertHandle prepared statement which does not match the type of
the SQL type of handle::deviceid.